+++ backend/Travel/Models/ThoiGian.cs
@@ -11,7 +11,7 @@ namespace Travel.Models
         public int TourId { get; set; }
         public Tour Tour { set; get; }
         public DateTime NgayDi { get; set; }
-        public int Gia { get; set; }
+        public decimal Gia { get; set; }
         public DateTime NgayVe { get; set; }
         public int SoLuongMax { get; set; }
         public int SoLuongDat { get; set; }
